refactor(storage): preserve store-owned file modes
Restore SettingsStore's historical 0666-and-umask behavior while keeping MCP and credential files explicitly 0600. Make fileMode required so the shared writer represents a real caller-owned policy and applies the effective mode before synchronization.

Move hardenDirectory and its coverage to stable-storage, update the two secret-store imports, and refresh the Windows test inventory.

Refs #4285

Generated-by: OpenAI Codex
